Refactor folder structure and fix comparison table text bug
