+++ client/src/components/App.tsx
@@ -19,14 +19,14 @@ import AdminPeakLists from './adminPanel/AdminPeakLists';
 import AdminRegions from './adminPanel/AdminRegions';
 import AdminStates from './adminPanel/AdminStates';
 import AdminUsers from './adminPanel/AdminUsers';
-import ComparePeakListPage from './comparePeakList';
-import CompareAllPeaksPage from './comparePeakList/CompareAllMountains';
-import MountainDetailPage from './mountainDetail';
-import PeakListDetailPage from './peakListDetail';
-import PeakListPage from './peakLists';
+import MountainDetailPage from './mountains/detail';
+import ComparePeakListPage from './peakLists/compare';
+import CompareAllPeaksPage from './peakLists/compare/CompareAllMountains';
+import PeakListDetailPage from './peakLists/detail';
+import PeakListPage from './peakLists/list';
 import Header from './sharedComponents/Header';
-import UserProfile from './userProfile';
-import ListUsersPage from './users';
+import UserProfile from './users/detail';
+import ListUsersPage from './users/list';
 
 const client = new ApolloClient();
 export const UserContext = React.createContext<User | null>(null);
