+++ client/src/components/peakLists/detail/MountainRow.tsx
@@ -1,26 +1,26 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
-import { mountainDetailLink } from '../../routing/Utils';
+import { mountainDetailLink } from '../../../routing/Utils';
 import {
   ButtonSecondary,
   lightBaseColor,
   lightBorderColor,
   semiBoldFontBoldWeight,
-} from '../../styling/styleUtils';
-import { CompletedMountain, PeakListVariants } from '../../types/graphQLTypes';
+} from '../../../styling/styleUtils';
+import { CompletedMountain, PeakListVariants } from '../../../types/graphQLTypes';
 import {
   failIfValidOrNonExhaustive,
   formatNumberWithCommas,
-} from '../../Utils';
-import { Months, Seasons } from '../../Utils';
+} from '../../../Utils';
+import { Months, Seasons } from '../../../Utils';
 import {
   formatDate,
   getFourSeasonCompletion,
   getGridCompletion,
   getStandardCompletion,
   getWinterCompletion,
-} from '../peakLists/Utils';
+} from '../Utils';
 import {
   MountainDatum,
 } from './index';
